Submitted and passes all tests
what would be nice is a way to only change part of the lines drawn when menu options are selected, also would be nice to add visual images to the dot, dash, solid lines and change the color of text on red, blue, black buttons. I would like to play with the sound libraries and have sounds play when certain options are selected.