Tested and works!
I would like to get it to display the name of the type of car, and only display the car if the gas tank has >0.4 gallons maybe use FLTK to graphically show cars going around a track. I also thought of using sounds to signal a set of five laps has elapsed. This would require using the libraries we downloaded in class but ran out of time.